♻️ (model/question) move ImageSpecs to new file
